webapi/tct-backgrounds-css3-tests/backgrounds/CSS3BG_background_example_equivalent.html

<!DOCTYPE html>
<!--
Copyright (c) 2013 Intel Corporation.

Redistribution and use in source and binary forms, with or without modification,
are permitted provided that the following conditions are met:

* Redistributions of works must retain the original copyright notice, this list
  of conditions and the following disclaimer.
* Redistributions in binary form must reproduce the original copyright notice,
  this list of conditions and the following disclaimer in the documentation
  and/or other materials provided with the distribution.
* Neither the name of Intel Corporation nor the names of its contributors
  may be used to endorse or promote products derived from this work without
  specific prior written permission.

THIS SOFTWARE IS PROVIDED BY INTEL CORPORATION "AS IS"
AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
ARE DISCLAIMED. IN NO EVENT SHALL INTEL CORPORATION BE LIABLE FOR ANY DIRECT,
INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING,
BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY
OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE,
EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.

Authors:
        haox.li <haox.li@intel.com>
        Cui, Jieqiong <jieqiongx.cui@intel.com>

-->

<html>
  <head>
    <title>CSS3 Backgrounds Test: CSS3BG_background_example_equivalent</title>
    <link rel="author" title="Intel" href="http://www.intel.com" />
    <link rel="help" href="http://www.w3.org/TR/2012/CR-css3-background-20120724/" />
    <meta name="flags" content="" />
    <meta name="assert" content="Check if the other individual properties are set to their initial values when only partial values for 'background' has been given" />
    <script src="../resources/testharness.js"></script>
    <script src="../resources/testharnessreport.js"></script>
    <script src="support/support.js"></script>
    <style type="text/css">
        #test{
            background: blue url("support/swatch-orange.png") 10% round fixed border-box;
        }
    </style>
  </head>
  <body>
    <div id="log"></div>
    <div id="test"></div>
    <script>
        test(function () {
            var div = document.querySelector("#test");
            var backgroundColor = GetCurrentStyle("backgroundColor");
            assert_equals(backgroundColor, "rgb(0, 0, 255)", "The element background-color value");

            var backgroundPosition = GetCurrentStyle("backgroundPosition");
            assert_equals(backgroundPosition, "10% 50%", "The element background-position value");

            var backgroundSize = GetCurrentStyle("backgroundSize");
            assert_equals(backgroundSize, "auto","The element background-size value");

            var backgroundRepeat = GetCurrentStyle("backgroundRepeat");
            assert_equals(backgroundRepeat, "round", "The element background-repeat value");

            var backgroundClip = GetCurrentStyle("backgroundClip");
            assert_equals(backgroundClip, "border-box","The element background-clip value");

            var backgroundOrigin = GetCurrentStyle("backgroundOrigin");
            assert_equals(backgroundOrigin, "border-box", "The element background-origin value");

            var backgroundAttachment = GetCurrentStyle("backgroundAttachment");
            assert_equals(backgroundAttachment, "fixed", "The element background-attachment value");

            var propvalue = GetCurrentStyle("backgroundImage");
            var prop = propvalue.indexOf("support/swatch-orange.png") != -1;
            assert_true(prop, "The element background-image should be url(support/swatch-orange.png)");
        }, document.title);
    </script>
  </body>
</html>
